Add token authentication to KASWebSocket connection
The KASWebSocket class has been updated to include a token for authentication during connection. The token is passed in upon initialization and is added to the header of the WebSocket request. The token is also sent as a WebSocket message on connection. The changes also affect KASWebSocket tests to accommodate the new parameter.